Move Req checkRssvVsSketch to datasketches/CrossCheckQuantiles
